+++ src/main/java/ufro/ignaciou/taller1_backend/models/Especie.java
@@ -17,7 +17,7 @@ public class Especie {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_especie")
-    private Long idEspecie;
+    private Integer idEspecie;
 
     @Column(name = "nombre")
     private Enum nombre;
